+++ uefi/src/proto/console/pointer/mod.rs
@@ -7,14 +7,14 @@ use core::mem::MaybeUninit;
 /// Provides information about a pointer device.
 #[repr(C)]
 #[unsafe_protocol("31878c87-0b75-11d5-9a4f-0090273fc14d")]
-pub struct Pointer<'boot> {
+pub struct Pointer {
     reset: extern "efiapi" fn(this: &mut Pointer, ext_verif: bool) -> Status,
     get_state: extern "efiapi" fn(this: &Pointer, state: *mut PointerState) -> Status,
     wait_for_input: Event,
-    mode: &'boot PointerMode,
+    mode: *const PointerMode,
 }
 
-impl<'boot> Pointer<'boot> {
+impl Pointer {
     /// Resets the pointer device hardware.
     ///
     /// The `extended_verification` parameter is used to request that UEFI
@@ -54,7 +54,7 @@ impl<'boot> Pointer<'boot> {
     /// Returns a reference to the pointer device information.
     #[must_use]
     pub const fn mode(&self) -> &PointerMode {
-        self.mode
+        unsafe { &*self.mode }
     }
 }