+++ CHANGELOG.md
@@ -20,7 +20,8 @@
 - `HandleBuffer` and `ProtocolsPerHandle` now implement `Deref`. The
   `HandleBuffer::handles` and `ProtocolsPerHandle::protocols` methods have been
   deprecated.
-- Removed `'boot` lifetime from the `Output` protocol.
+- Removed `'boot` lifetime from the `GraphicsOutput`, `Output`, `Pointer`, and
+  `Serial` protocols.
 - The generic type `Data` of `uefi::Error<Data: Debug>` doesn't need to be
   `Display` to be compatible with `core::error::Error`. Note that the error
   Trait requires the `unstable` feature.